Fix (for real) keyboard navigation
The issue was that each row would have an index assigned when the row is created. Then, when the enter key is pressed, the row is picked by finding a row with index+1. The issue is if a new row is created or deleted, these indexes aren't updated, meaning that an index might be missing or duplicated. While these indexes could be fixed, I opted for the approach of assigning a unique ID (just an increasing number) and then passing that into the keyboard bind. Then, we look for the index of the row with the matching ID.